contributing: add PR validation checklist
Turn the contributing section in a checklist. This checklist should
be used when deciding on whether a PR should be accepted as is or if
modifications should be asked. If an answer to one of the questions
on the checklist is 'no', the branch should be modified accordingly.

This commit also specify that urls should be declared using the
appropriate templatetag and that strings should be internationalised.