Add the possibility to attach an existing member to the client.
Changes made:
- Add a route and a view for making ajax request in order to
find the existing members.
- Update the templates to add the new input to search existing
members and the JavaScript code needed.
- Update the view ClientWizard to add the existing member to the
client when needed, also split the long save method into several
logic methods that are reused in the save method.

- * Important change * : Right now I have decided that if the
client information and the payment information is the same
(First Name and Last Name) only one member is created in the
DataBase, because if not we end with lots of members that basically
have the same information. Is the same with the payment information
and the referent information, maybe the person referencing the
client is the same that will pay for it, and right now we create
two differents members in the DataBase. I'm not sure how this should
be handled and I will be happy to change this into the right way.

Ref: Issue # 223