Fix bug in the form validation and add more tests.
The street, city and postal_code fields are not required if payment
information is an existing member with a valid address. We can use
the address already in the system, not need to add a new one, more
when right now a member can only have an address, something that we
should change? probably!