Replace dataload and dataexec with initial_data fixture, original idea @cormier
Cormier has proposed to use a fixture instead of dataexec and dataload, and
that is a good idea to me. But I prefer not to delete the dataexec and dataload
files, because the fixtures created was a dump of the database after populate
the database with dataexec and dataload, so it will be easier in the future
to made modifications in dataload and use it to generate a new version of
the fixture initial_data.json.

Ref: issue #357  pr #359