Fix issue #502 Geolocation is broken in the client creation form.
Split the multidatespicker javascript files into a different files to include
only in the page for create orders in batch.
Change the versions of semantic-ui, jquery, jquery-ui to avoid some incompatibilities.
Remove generated assets from repo to avoid anyone change the generated files.
Install npm and gulp in the web image because we need gulp to generate the assets.

Ref: issue-502