Fix collect static warnings.
Rename the destination folder for the static assets generated by gulp inside
sous_chef app,  the static files finders will look inside each app for an
static folder and the first files found will be used. This is causing that the
last compiled files are ignored by the static files command when running
the project using the production settings.
https://docs.djangoproject.com/en/1.11/ref/settings/#std:setting-STATICFILES_FINDERS

Fix grid in the user edit and create page.

Ref: issue #782